created stateless widgets for rating,genre and manga details
