+++ lib/screens/mangafox/manga_details.dart
@@ -1,12 +1,11 @@
 import 'package:bot_toast/bot_toast.dart';
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:expandable/expandable.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:manga_reader/constants/assets.dart';
+import 'package:manga_reader/custom/widgets/manga_details.dart';
 import 'package:manga_reader/custom/widgets/scale_route_transition.dart';
 import 'package:manga_reader/models/manga.dart';
 import 'package:manga_reader/screens/reader_loader.dart';
@@ -46,7 +45,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
         if (r.url == null || r.url != widget.manga.url) {
           context
               .read<MangafoxDetailsProvider>()
-              .getMangaDetails(Assets.lelscanCatalogName, widget.manga);
+              .getMangaDetails(Assets.mangafoxCatalogName, widget.manga);
         }
       });
       context.read<MangafoxChapterProvider>().mangaChapters.fold((l) => null, (r) {
@@ -55,7 +54,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
             widget.manga != context.read<MangafoxChapterProvider>().currentManga) {
           context
               .read<MangafoxChapterProvider>()
-              .getChapters(Assets.lelscanCatalogName, widget.manga);
+              .getChapters(Assets.mangafoxCatalogName, widget.manga);
         }
       });
     });
@@ -100,7 +99,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
                             context
                                 .read<ActionProvider>()
                                 .downloadMultipleChapters(
-                                Assets.lelscanCatalogName,
+                                Assets.mangafoxCatalogName,
                                 widget.manga.title,
                                 MediaQuery.of(context).size);
                           }),
@@ -131,6 +130,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
               : FloatingActionButton(
             backgroundColor: Colors.cyan,
             child: Icon(Icons.play_arrow, color: Colors.white),
+            onPressed: (){},
           ),
           appBar: context.watch<ActionProvider>().selectedItems.isEmpty
               ? AppBar(
@@ -223,6 +223,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
                           ),
                         );
                       }, (mangaDetails) {
+                        print(mangaDetails.rating);
                         return mangaDetails.detailsFetched != true
                             ? Container(
                           child: Center(
@@ -239,7 +240,7 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
                                     SizeConfig.blockSizeVertical *
                                         3),
                                 child:
-                                _buildMangaDetails(mangaDetails),
+                                MangaDetails(manga:mangaDetails),
                               ),
                               SizedBox(
                                 height:
@@ -602,206 +603,6 @@ class _MangafoxDetailState extends State<MangafoxDetail> {
     );
   }
 
-  Widget _buildMangaDetails(Manga manga) {
-    return Padding(
-      padding: EdgeInsets.symmetric(
-          horizontal: SizeConfig.blockSizeHorizontal * 2.5),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Container(
-            width: SizeConfig.screenWidth / 2.2,
-            child: CachedNetworkImage(
-              imageUrl: widget.manga.thumbnailUrl,
-              width: double.infinity,
-              height: 250,
-              errorWidget: (context, text, data) {
-                return Image.asset(
-                  Assets.errorImage,
-                  width: double.infinity,
-                  height: 250,
-                );
-              },
-              fit: BoxFit.fill,
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.only(left: SizeConfig.blockSizeVertical),
-            child: Container(
-              width: SizeConfig.screenWidth / 2.2,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    manga.title,
-                    overflow: TextOverflow.clip,
-                    style: TextStyle(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                        fontSize: 17.0),
-                  ),
-                  SizedBox(
-                    height: SizeConfig.blockSizeVertical * 2,
-                  ),
-                  RichText(
-                    textAlign: TextAlign.center,
-                    text: TextSpan(
-                        text: 'de ',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                        children: <TextSpan>[
-                          TextSpan(
-                            text: manga.author,
-                            style: TextStyle(
-                              color: Colors.cyan,
-                            ),
-                          )
-                        ]),
-                  ),
-                  SizedBox(
-                    height: SizeConfig.blockSizeVertical * 2,
-                  ),
-                  RichText(
-                    textAlign: TextAlign.center,
-                    text: TextSpan(
-                        text: 'dessiné par ',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                        children: <TextSpan>[
-                          TextSpan(
-                            text:
-                            manga.artist.isEmpty ? "inconnu" : manga.artist,
-                            style: TextStyle(
-                              color: Colors.cyan,
-                            ),
-                          )
-                        ]),
-                  ),
-                  SizedBox(
-                    height: SizeConfig.blockSizeVertical * 2,
-                  ),
-                  _buildGenres(manga.genre),
-                  SizedBox(
-                    height: SizeConfig.blockSizeVertical * 2,
-                  ),
-                  RichText(
-                    textAlign: TextAlign.center,
-                    text: TextSpan(
-                        text: 'statut ',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                        children: <TextSpan>[
-                          TextSpan(
-                            text: manga.status,
-                            style: TextStyle(
-                              color: Colors.cyan,
-                            ),
-                          )
-                        ]),
-                  ),
-                  SizedBox(
-                    height: SizeConfig.blockSizeVertical * 2,
-                  ),
-                  RichText(
-                    textAlign: TextAlign.center,
-                    text: TextSpan(
-                        text: 'source ',
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
-                        children: <TextSpan>[
-                          TextSpan(
-                            text: manga.catalog,
-                            style: TextStyle(
-                              color: Colors.cyan,
-                            ),
-                          )
-                        ]),
-                  ),
-                  _buildRating(manga.rating)
-                ],
-              ),
-            ),
-          )
-        ],
-      ),
-    );
-  }
-
-  Widget _buildRating(String rate) {
-    print(rate);
-    if (rate != null) {
-      var data = rate.split(" ");
-      var mark = data[0].split("/");
-      return Container(
-        width: SizeConfig.screenWidth / 1.5,
-        child: RatingBar.builder(
-          initialRating: double.parse(mark[0]),
-          tapOnlyMode: true,
-          minRating: 0,
-          updateOnDrag: false,
-          unratedColor: Colors.amber.withOpacity(0.5),
-          direction: Axis.horizontal,
-          allowHalfRating: true,
-          itemCount: int.parse(mark[1]),
-          itemBuilder: (context, _) => Container(
-            width: SizeConfig.blockSizeHorizontal * 4,
-            child: Icon(
-              Icons.star,
-              size: 8,
-              color: Colors.amber,
-            ),
-          ),
-        ),
-      );
-    } else {
-      return Container(
-        width: SizeConfig.screenWidth / 1.5,
-        child: RatingBar.builder(
-          initialRating: 0,
-          tapOnlyMode: true,
-          updateOnDrag: false,
-          unratedColor: Colors.amber.withOpacity(0.5),
-          minRating: 0,
-          direction: Axis.horizontal,
-          allowHalfRating: true,
-          itemCount: 5,
-          itemBuilder: (context, _) => Container(
-            width: SizeConfig.blockSizeHorizontal * 4,
-            child: Icon(
-              Icons.star,
-              size: 8,
-              color: Colors.amber,
-            ),
-          ),
-        ),
-      );
-    }
-  }
-
-  Widget _buildGenres(String genres) {
-    final data = genres.split(",");
-    return Wrap(
-        runSpacing: 5.0,
-        spacing: 5.0,
-        children: List.generate(data.length, (index) {
-          return Container(
-            padding: EdgeInsets.symmetric(horizontal: 5.0),
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.cyan),
-              borderRadius: BorderRadius.all(Radius.circular(25)),
-            ),
-            child: Text(
-              data[index],
-              style: TextStyle(color: Colors.cyan),
-            ),
-          );
-        }));
-  }
 
   Widget _builButtons() {
     return Padding(
